Finish timer function
I have introduced a timer to handle expired connections.
Additionally, I have fixed the socket settings. Now,
the socket connection is set to 'keep-alive' instead of
'close', allowing the socket to be used for multiple
requests instead of just one.

Now, when there are no new requests on the connection
after the default timeout, it will be disconnected.
The default timeout is set to 10.